Great push of test quest
A lot of changings in basic template to complete test of CSSSR job.
Background changed, different logos and other blocks and elements created or modified (if they are suitable for my work).
I need to add other elements, change position to look the same as in .psd file and code good mobile look of this page.